# api/openapi.yaml
openapi: 3.0.3
info:
  title: Poydem API
  version: 1.0.0
  license:
    name: Proprietary
  description: |
    Final frontend/backend contract for Poydem MVP API v1.

    All enum values use lower-case except stable error and system-reason codes,
    which use UPPER_SNAKE_CASE. Every JSON response uses UTF-8. Every response
    includes X-Request-ID. Timestamps are RFC 3339 date-time values with an offset.

    Access tokens are sent as Bearer JWTs. Refresh tokens exist only in an HttpOnly
    cookie and are never returned in JSON or URLs. Lists documented with pagination
    return `items: []` and `totalPages: 0` when empty. Unexpected server failures
    return HTTP 500 with error.code `INTERNAL_ERROR` in the common error envelope.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /users/me/companies:
    get:
      operationId: listMyCompanies
      tags: [Users]
      summary: My companies
      description: Returns current, non-deleted company memberships. A blocked company remains visible to its current members and owner with status blocked.
      x-error-codes:
        '400': [VALIDATION_ERROR]
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED]
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Companies
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CompanyList'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /users/me/applications:
    get:
      operationId: listMyApplications
      tags: [Users]
      summary: My applications
      description: Paginated application history across all statuses, including old rejected and cancelled records.
      x-error-codes:
        '400': [VALIDATION_ERROR]
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED]
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Applications
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApplicationList'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /events/{eventId}/companies:
    get:
      operationId: listEventCompanies
      tags: [Events]
      security:
        - {}
        - bearerAuth: []
      summary: Event companies
      x-error-codes:
        '400': [VALIDATION_ERROR]
        '404': [EVENT_NOT_FOUND]
      parameters:
        - {$ref: '#/components/parameters/EventId'}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Companies
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CompanyList'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
    post:
      operationId: createCompany
      tags: [Companies]
      summary: Create company
      description: Atomically creates the company, adds the creator as owner/member, and records their event participation. The owner counts toward maxMembers.
      x-error-codes:
        '400': [VALIDATION_ERROR]
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, PROFILE_INCOMPLETE]
        '404': [EVENT_NOT_FOUND]
        '409': [EVENT_NOT_AVAILABLE, ALREADY_IN_EVENT_COMPANY]
      parameters: [{$ref: '#/components/parameters/EventId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CompanyInput'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Company'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /companies/{companyId}:
    get:
      operationId: getCompany
      tags: [Companies]
      security:
        - {}
        - bearerAuth: []
      summary: Company details
      description: Deleted companies are never returned. Blocked companies are visible only to current members, the owner, and admins.
      x-error-codes:
        '404': [COMPANY_NOT_FOUND]
      parameters: [{$ref: '#/components/parameters/CompanyId'}]
      responses:
        '200':
          description: Company
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Company'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      operationId: updateCompany
      tags: [Companies]
      summary: Update company
      x-error-codes:
        '400': [VALIDATION_ERROR]
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, NOT_COMPANY_OWNER]
        '404': [COMPANY_NOT_FOUND]
        '409': [COMPANY_BLOCKED, EVENT_NOT_AVAILABLE, COMPANY_CAPACITY_BELOW_MEMBERS]
      parameters: [{$ref: '#/components/parameters/CompanyId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateCompany'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Company'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
    delete:
      operationId: deleteCompany
      tags: [Companies]
      summary: Delete company
      description: Soft-deletes the company, cancels pending applications, and removes event participation linked to this company in one transaction.
      x-error-codes:
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, NOT_COMPANY_OWNER]
        '404': [COMPANY_NOT_FOUND]
      parameters: [{$ref: '#/components/parameters/CompanyId'}]
      responses:
        '204': {description: Deleted}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /companies/{companyId}/members:
    get:
      operationId: listCompanyMembers
      tags: [Companies]
      security:
        - {}
        - bearerAuth: []
      summary: Company members
      x-error-codes:
        '400': [VALIDATION_ERROR]
        '404': [COMPANY_NOT_FOUND]
      parameters:
        - {$ref: '#/components/parameters/CompanyId'}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserShortList'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /companies/{companyId}/members/me:
    delete:
      operationId: leaveCompany
      tags: [Companies]
      summary: Leave company
      description: Removes both company membership and event participation. It never converts the user to solo participation.
      x-error-codes:
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED]
        '404': [COMPANY_NOT_FOUND]
        '409': [OWNER_CANNOT_LEAVE, NOT_COMPANY_MEMBER]
      parameters: [{$ref: '#/components/parameters/CompanyId'}]
      responses:
        '204': {description: Left}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /companies/{companyId}/members/{userId}:
    delete:
      operationId: removeCompanyMember
      tags: [Companies]
      summary: Remove member
      description: Removes both company membership and event participation in one transaction.
      x-error-codes:
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, NOT_COMPANY_OWNER]
        '404': [COMPANY_NOT_FOUND, USER_NOT_FOUND]
        '409': [OWNER_CANNOT_BE_REMOVED, NOT_COMPANY_MEMBER]
      parameters:
        - {$ref: '#/components/parameters/CompanyId'}
        - {$ref: '#/components/parameters/UserId'}
      responses:
        '204': {description: Removed}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /companies/{companyId}/join:
    post:
      operationId: joinOpenCompany
      tags: [Companies]
      summary: Join open company
      description: Atomically joins an open company and changes existing solo participation to company participation when needed.
      x-error-codes:
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, PROFILE_INCOMPLETE]
        '404': [COMPANY_NOT_FOUND]
        '409': [COMPANY_FULL, COMPANY_CLOSED, EVENT_NOT_AVAILABLE, ALREADY_COMPANY_MEMBER, ALREADY_IN_EVENT_COMPANY]
      parameters: [{$ref: '#/components/parameters/CompanyId'}]
      responses:
        '201': {description: Joined}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /companies/{companyId}/close:
    post:
      operationId: closeCompanyRecruitment
      tags: [Companies]
      summary: Close recruitment
      description: Prevents new joins and applications. Existing pending applications may still be approved or rejected.
      x-error-codes:
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, NOT_COMPANY_OWNER]
        '404': [COMPANY_NOT_FOUND]
        '409': [COMPANY_BLOCKED]
      parameters: [{$ref: '#/components/parameters/CompanyId'}]
      responses:
        '200':
          description: Closed
          content: {application/json: {schema: {$ref: '#/components/schemas/Company'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly cookie managed by the browser; JavaScript cannot read it
  parameters:
...
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
  responses:
    BadRequest:
      description: Bad request
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Unauthorized:
      description: Unauthorized
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Forbidden:
      description: Forbidden
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
  schemas:
...
    CompanyList:
      type: object
      required: [items, pagination]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Company'}}
        pagination: {$ref: '#/components/schemas/Pagination'}
    ApplicationList:
      type: object
      required: [items, pagination]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Application'}}
        pagination: {$ref: '#/components/schemas/Pagination'}
...